Added location suggest to search page
